requirements.txt
absl-py==1.2.0
accelerate==0.24.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==3.6.1
apache-beam==2.45.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
attr==0.3.1
attrs==21.4.0
autopep8==1.7.0
avro-python3==1.10.2
azure-core==1.31.0
azure-storage-blob==12.23.1
Babel==2.10.3
backcall==0.2.0
bcrypt==4.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blis==0.7.11
bokeh==2.4.3
boto==2.49.0
boto3==1.35.49
botocore==1.35.49
boxing==0.1.4
cachetools==5.2.0
catalogue==2.0.10
certifi==2022.6.15
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudpathlib==0.18.1
cloudpickle==2.2.1
colorama==0.4.6
colorgram.py==1.2.0
comm==0.1.2
confection==0.1.5
contextlib2==21.6.0
contourpy==1.1.1
crcmod==1.7
cryptography==43.0.3
cycler==0.11.0
cymem==2.0.8
Cython==0.29.33
databases==0.9.0
db-dtypes==1.0.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.1.1
distlib==0.3.6
distro==1.9.0
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
dm-tree==0.1.8
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
ecdsa==0.19.0
email_validator==2.2.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
et-xmlfile==1.1.0
etils==1.0.0
executing==0.8.3
expiringdict==1.2.2
fastapi==0.115.5
fastapi-cli==0.0.5
fastapi-jwt-auth==0.5.0
fastavro==1.7.1
fasteners==0.18
fastjsonschema==2.15.3
filelock==3.9.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
flatbuffers==23.5.26
fonttools==4.33.3
fsspec==2023.10.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.21.0
google-api-python-client==2.79.0
google-auth==2.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.14.2
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-cloud-storage==2.18.2
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpc-google-iam-v1==0.13.1
grpcio==1.60.0
grpcio-status==1.47.0
h11==0.14.0
h5py==3.7.0
hdfs==2.7.0
htmlmin==0.1.12
httpcore==1.0.6
httplib2==0.20.4
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.19.4
humansignal-drf-yasg==1.21.9
idna==3.3
ijson==3.2.3
imageio==2.25.1
immutabledict==2.2.3
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
inquirerpy==0.3.4
install==1.3.5
invisible-watermark==0.2.0
ipykernel==6.21.2
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.7.2
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.1
Jinja2==3.1.2
jiter==0.6.1
jmespath==1.0.1
joblib==1.3.2
jovian==0.2.41
json5==0.9.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.6.1
jupyter-server==1.18.1
jupyter_core==5.2.0
jupyterlab==3.4.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.15.1
kaggle==1.5.12
keras==2.15.0
Keras-Preprocessing==1.1.2
keyring==23.13.1
kiwisolver==1.4.2
label-studio==1.13.1
label-studio-converter==0.0.59
label-studio-sdk==1.0.5
label-studio-tools==0.0.4
labelImg==1.8.6
langcodes==3.4.0
language_data==1.2.0
launchdarkly-server-sdk==8.2.1
libclang==14.0.6
lockfile==0.12.2
lvis==0.5.3
lxml==4.9.2
marisa-trie==1.2.0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mglearn==0.1.9
mistune==0.8.4
ml-dtypes==0.2.0
more-itertools==9.0.0
mpmath==1.3.0
murmurhash==1.0.10
nbclassic==0.4.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==3.2.1
nltk==3.9.1
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
objsize==0.6.1
openai==1.52.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
opendatasets==0.1.22
openpyxl==3.0.10
opentelemetry-api==1.27.0
opt-einsum==3.3.0
ordered-set==4.0.2
orjson==3.8.6
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathlib==1.0.1
pfzy==0.3.4
pickleshare==0.7.5
pillow==11.0.0
pkginfo==1.9.6
platformdirs==3.0.0
portalocker==2.7.0
preshed==3.0.9
prettytable==3.4.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.37
proto-plus==1.25.0
protobuf==3.20.3
psutil==5.9.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.19
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.18.4
pydot==1.4.2
pygal==3.0.0
pygal-maps-world==1.0.2
pygalib==0.1.0
pygame==2.1.2
Pygments==2.14.0
PyJWT==1.7.1
pymongo==3.13.0
PyMuPDF==1.24.14
pyparsing==2.4.7
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyRFC3339==1.1
pyrsistent==0.18.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.4
python-multipart==0.0.17
python-slugify==6.1.2
pythonds==1.2.1
pytz==2022.1
PyWavelets==1.5.0
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.5
PyYAML==6.0.2
pyzmq==22.3.0
qtconsole==5.4.0
QtPy==2.3.0
readme-renderer==37.3
redis==3.5.3
regex==2022.10.31
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.3
sacrebleu==2.2.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.9.0
seaborn==0.13.0
semver==3.0.2
Send2Trash==1.8.0
sentencepiece==0.1.97
sentry-sdk==2.17.0
seqeval==1.2.2
shellingham==1.5.4
six==1.16.0
sklearn==0.0
smart-open==7.0.4
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.7.5
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.2.0
starlette==0.41.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-addons==0.19.0
tensorflow-datasets==4.8.2
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.12.0
tensorflow-intel==2.15.0
tensorflow-io==0.30.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-metadata==1.12.0
tensorflow-model-optimization==0.7.3
tensorflow-object-detection-api==0.1.1
tensorflow-text==2.10.0
termcolor==1.1.0
terminado==0.13.3
text-unidecode==1.3
tf-models-official==2.10.1
tf-slim==1.1.0
thinc==8.2.4
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.15.0
toml==0.10.2
torch==2.1.1
tornado==6.1
tqdm==4.64.1
traitlets==5.9.0
transformers==4.35.2
twine==4.0.2
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ua-parser==0.18.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
uuid==1.30
uvicorn==0.32.1
virtualenv==20.19.0
wasabi==1.1.3
watchfiles==1.0.0
wcwidth==0.2.5
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.4.1
websockets==14.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
WTForms==3.0.1
wurlitzer==3.0.3
xgboost==1.6.2
xmljson==0.2.1
zipp==3.14.0
zstandard==0.20.0
